Added option to enqueue a mirror job after harvest. Closes #11
Top level path is specified in `settings.actions.harvest.copy_base_path` on dequeue worker.
Rake task for harvesting has an extra option `copy_on_success`, which can be `no_copy` or `copy_on_success`.
Enqueing the copy task is part of the harvest - the harvested file will be set to aborted if the copy enqeue is not successful.