A suite of redis memory optimizations
Fixes #40. Fixes #41. Closes #31.

- Monkey patches resque:status so that TTLs are only set on finished
status
- Ensures resque:status's expire_in option is set for both work and web
startup scenarios
- Forces action implementers to override resque:status' default name
implentation to prevent double encoding of payloads
- Added friendly name methods to all actions
- Introduces the concept of partial payloads (splitting payloads into
variant and invariant parts) to reduce memory usage for large sets of
similar jobs
- It adapts the Analysis action (only) to use partial payloads. Users of
`action_enqueue` need to provide a group key to take advantage of the
functionality.
- We introduce our own Redis communicator for access to Redis outside of
Resque's scope
- It drys up the `config.rb` a bit.
- We add a template action as clean base (mainly done to help @atruskie
learn what is necessary for an Action to implement and what among the
other examples was extraneous to a minimal example).

:grinning: