Completed menu UI changes
Work for #275 .

There are lots of changes.

There are still links that need to be cleaned up in the admin section.

There likely is a better way to do a lot of this too. In paritcular I'm
thinking some kind of MVVM methodolgy (where view specific variables can
be defined in code for entire folders). Even so, a lot of helper
functions have been defined that will ensure consistency of menu links.

However we need to ship, so smaller changes/refactors/fixes will be done
at a later time.