Created an alphabetical pager
This pager exposes a an application controller (which renders a UI
partial) along with an active record extension that makes it easy to
filter a textual field based on lexiographical categories. Note: it
technically is not  pager because there a no fixed size pages of results
(and commonly there will be no items on some pages).

It's logic is limited for now because it currently fulfills all
functionality required. However I see it being useful in the REST API as
a filter function. I also expect it could be further expanded to get
totals and to also dynamically remove irrelevant pages.

This was done for #322