Application controller checks that post params is not empty for update and create
If there are no params, determines whether the request body was empty or if it just was not parsed properly, and responds with approprtiate error code depending on the situation.
If the body was empty, includes an error link to the new action.
If the body was not empty, responds with unsupported_media_type.