Updated dockerfile so that binstubs are used
With binstubs it is much harder to make a confusing and time wasting mistake because bundler binstubs automatically act like they're being used inside of bundle exec.

I've also made the test fail if they're not executed with bundle and I've foreced the tests to run in RAILS_ENV=development mode.