Fixes all model tests
Mainly involved adding :optional to associations that were optional (since associations are now required by default in rail 6+)

Also had to override more of the big decimal json serialization so it would actually take effect when serializing notes to the database.

Defined helper methods to refer to the default dataset as well and ensured the helpers were used in the seeds and tests.

-> { uniq}  is deprecated and needed to be replaces with distinct.

the activesupport parameterize method requires named arguments to work now.