Fix saved search tests, upgrade stored_query to JSON
Partially adresses #467.

The now deprecated `serialize` constrtuct that used to be applied to models that store hashes in a text field is *now* taken into account when AREL constructs queries. The result was a double-encoded JSON string used a value to search for in the database which would obviously never match anything!

Made the following changes to adress this:
- simplified strong params
- migrated stored_query to jsonb
- allowed our API code to accept a Hash as a primitive value to compare against in queries, IFF the type of the column of the attribute is JSON.
- and adjusted the acceptance test so that a true hash is passed rather than a JSON encoded string
Also added patches to AREL for retrieving the type of an attribute.

Also canonicalised our is_json postgres function.