Redesigned timezone normalization
Modelled the timezone helper as a module so its behaviour was more  testable and uniform.

Functional changes include: APIs now accept IANA identifiers and friendly identifiers (as opposed to only the non-standard friendly identifiers), IANA identifiers are now stored in the database, and friendly identifiers are converted when read or assigned to the database.

All of this is now covered by comprehensive unit tests.

Also fixed a bug with haml templates escaping JSON strings.