Address PR review
Also fixes authorization check and properly configures CMS auth.

Also attempts to deal with intermittently failing test - depending on the order the test is run the Resque.logger has either been properly initialized or not. In the latter case our guard clause that checks the resque logger is correct fails (as  a false positive). Instead I changed the call of the run_web method in the  resque_status_spec test to use the correct logger.

Also fixed a bug that saw state shared between the DSL used in permissions tests. The new attributes are isolated in each group and even support inheritable overrides - which is what I really wanted.