Adds an upload service
This upload service allows for users to upload audio without having server ssh credentials.

This is part 1 of a larger set of changes. None of this functionality is exposed at the moment. Rather this change set just adds the service and makes sure it is usable.

The upload server itself is essentially a microservice. It is a go package that allows for sftp, webdav, and ftp uploading and has a REST based admin API.

We use the REST API to provision (and chroot) upload users that each harvest process will be able to use. Access is scoped, and temporary.

Note all the code in the lib/gems/sftpgo_generated_client directory was generated by the openapi code generation tool. The command used is stored in provision/sftpgo_schema_generation.sh

I also added `turnip` - a gherkin/cucumber implementation for rspec. Its BDD style makes it a litte easier to test complex services.

Also refactored Dockerfile with view to avoid deploy builds.

BREAKING CHANGE