Simplifies upload service API client
Removes the generated OpenAPI code. The code was verbose, repititive, and also had errors due to some faults in the API specification (and the code  generator had bugs in interpreting the spec as well).

Replaced with a hand-written client that is smaller and neater. It does less, but we currently do not need more than is implemented so that is another win.