Remove asset compilation from dockerfile
There's just no easy way to compile assets in a docker environment.

The precompiler boots rails, which loads initializers which depends on a database connection (which is not available). Even mocking the database (with a NullDatabase, or  sqlite) doesn't work because our settings files are intentionally missing values for production environments (for safety reasons).