Attempt to simplify the docker file
The old dockerfile had two targets (dev & prod). This lead to a lot of repeated logic for relatively few customisations.

It also made build times slow because bundle install was repeated several times.

I suspect it was also bloating image sizes as well.

This change removes the targets and moves dev-logic into entrypoint.sh that runs when an appropriate environment is detected.