Fix bug in upload service communicator
It was not correctly using https in production because URI::HTTP will only ever produce a HTTP url, even if you specify the scheme as something else.

It also does not throw an error when you try to use a different scheme; hence the silent failure.

Added tests for this case as well.