Change app initialization
Now, a custom config file will actually be loaded. This was important for making workers actually work in prod!

Also modified the PauseDequeueForTests module so that the constant is not referred anywhere else. Zeitwerk was trying to load the constant when we were doing our check to make sure the module was not integrated.....  which of course loaded the module.
If the module is was missing, then zeitwerk would fail when it could not load the constant.