Formalised RSpec metdata state storage
Storing state in helpers/extentions in RSpec is a minor nightmare because describe blocks are sub-classes of parent describe blocks.

That means a helper that needs state stored has to do wild incantations for class variables and deal with inheritance, and deal with access when the example group is eventually instantiated.

RSpec has metdata to solve this issue, but access to that can also be very confusing.

I've added a module that defines wrapper methods for RSpec's metadata for every context I could think of - in the describe block, in before/after blocks, and in examples.

This should make state manipulation far less error prone. It also comes with a trace function so you can see at which scope a metadata entry was maniuplated!

This is used in the resque_helpers extension currently but can also be used elsewhere.