Decouples spectogram generation from audio generation
When generating media a spectrogram first needs to cut out the audio it needs. If there's a audio job running for the same segment at the same time we get race conditions.

This change makes the spectorgram generation independent of the audio segment (they don't use the same file). This technically makes the service slower (as cut audio won't be used from the cache as frequently). Although not measured, I'd estimate the time it takes to deal with an exception and send an error notification is far longer though.

So net win in terms of performance and reliability.

Also packages in reliability changes that were found to be needed while testing out these race conditions.

Also adds `no_header` to ffmpeg invocations to significantly remove noise from ffmpeg error messages.

Fixes #527