Fix: split upload.host into public and admin hosts
Production networking is more complex than dev envirionment. Production hosts are load balanced for security reasons so the public host advertised needs to be different from the host used to adminstrate our upload service.

Also wraps harvest#update in a row-lock transaction.

Also makes the app more resiliant to missing directories on start up - this is necessary for when storage systems are in maintenance mode.