Fix race condition in harvests
When wrapping the entire enquue process (for metadata_extraction/process stages) in a single transaction, the HarvestJobs that are enqueued first can finish before the transaction is committed. When the transaction is comitted then stale status data is written to the harvest_items table.

This is no longer the case; we removed two serparate transaction blocks.

Fixes #590