Fix harvest item validations failing against items in another harvest
We can't really control what's happening in another harvest. Instead we make sure there are no overlaps/duplicates in the current harvest and the in audio_recordings. When the final ingest happens we still fall back to whoever wins first.

The alternative - make sure in progress harvests don't conflict with other in progress harvests  - is a nigtmare to solve, especially from a UI perspective. We also believe it to be the rarest of use cases.

Fixes #605