Makes EMU unfixable fixes idempotent
Fixes #609.

EMU now no longer renames files that are unfixable. This makes it easier for us to consistently server validations.

Previously if a file was renamed (e.g. suffixed with `.error_empty`) it would not fail validations on the second run through and instead other validations might raise (e.g. error bad extension) or no validations would raise which would result in even more errors.

Also added a version field the harvest info struct. This makes it a little easier to track the multiple invocations of emu across runs.

Also removed an invalid fix result constant from emu (affected is not a result of a fix, only a check).