Refactors site location jitter
Site locations were rendering as much as 2km away from the true points.

We've refactored jittering so it now uses a simpler method of calculation, and is now also stable for any given set of coordinates. A stable jitter is more secure (you can't brute force average jittered samples to get a true location) and is more user friendly.

We dropped the jit range down to 333m (at the equator). Max displacement will now be 150ish m, with an exclusion zone of 30m).

Fixes #620