Allows jobs to be cancelled efficiently
We now maintain two implementations of cancelling a job on the remote queue. The state machine version is hte original and approximately 10x slower than the batch version.

There are some tradeoffs, notably, we no longer collect performance metrics for jobs that were cancelled and we can do no longer just cancel the subset of jobs that need to be cancelled. Instead everything on the remote queue, for a particular job, is cancelled. Functionally this is different from serial cancellation, but cancellation is only ever done in the context of cancelling every item anyway (we never needed to cancel just one job item) so we expect this change in behaviour ot be ok.