Fix remote status job not working in production
I had accidentally put the `check_item` call in a logger payload generation block. The intention was to use `debug_measure` which runs the code no matter the log level. But plane old `debug` uses the block for payload values and isn't evaluated when the log level is not DEBUG - which it isn't on production.

In the course of investigating this I also fixed a few other things.

The status checking job now limits itself to working on smaller batches.

The concurrency module for active job no longer makes count keys in redis for jobs that don't have a concurrency limit.

I also removed extraneous double colons in redis keys. It's only conventional to use a single colon for key token separation but it makes some developer tools (like redis insights) easier to use.