Improves script validation for settings/config
Now validates that is any of the script *settings* attributes are set, that they are all set.

It also validates that the command doesn't contain the config template tokens if the settings are not provided.

Finally, it makes sure the communicator will template an empty config file. It used to not template empty an empty config, but an empty config is valid.