Adds a prelude to analysis scripts
Fixes #766.

Adds the ability to configure a script prelude for analysis job scripts.

This kind of fixes #766: the issue was to do with the host environment and so actually changing the server wasn't appropriate. But allowing us to modify the host environment before the rest of a job runs is a good solution.

Also removes the cancel_jobs function. It was not reliable when trying to delete jobs that were exiting.