Up curl status retry count for batch jobs
The retry count of `3` resulted in only `7` seconds window for retries, which is too short for our services to recover.

Upping the retry count to `10` will let us retry over a window of `1023` seconds (about 17 min).

I stuck with `--retry` and did not use `--retry-after` because retry after, when set, disables the backoff algorithm in curl. Backoff is desirable.

Fixes #782